agregando hasta el ejercicio 13 guia 1

+++ Unidad1/ejer06/main.c
@@ -4,11 +4,11 @@
 
 int main()
 {
-    float x, tol;
+    double x, tol;
     printf("Ingrese un valor de X para la funcion sen(X): ");
-    scanf("%f", &x);
+    scanf("%lf", &x);
     printf("Ingrese un valor de tolerancia (valor entre 0 y 1): ");
-    scanf("%f", &tol);
+    scanf("%lf", &tol);
     printf("\nEl valor de sen(%.2f) es %.3f", x, senX(x, tol));
     return 0;
 }
